// src/interfaces/finance/accounts/transactions/chequePaymentTransaction.interface.ts
export interface ITrAcInvdetail {
  c_bal_amt_org?: number;
  inv_amt?: number;
  IsDeletable?: boolean;
  company_code: string;
  doc_type: string;
  doc_no: number;
  serial_no: number;
  dtl_sr_no: number;
  doc_date: Date;
  ac_code: string;
  inv_no: string;
  inv_date: Date;
  due_date: Date;
  chq_no: string;
  chq_date: Date;
  chq_bank: string;
  amount: number;
  lcur_amount: number;
  sign_ind: number;
  curr_code: string;
  ex_rate: number;
  ex_rate_origin: number;
  curr_code_origin: string;
  amount_origin: number;
  indicator_origin: string;
  div_code: string;
  oracle_upload: string;
  oracle_dt: Date;
  created_by: string;
  created_at?: Date;
  updated_by: string;
  updated_at?: Date;
}